Added security check to checklist_manager

+++ brighterchecklist/checklist_manager/views.py
@@ -1,17 +1,13 @@
-from django.shortcuts import render, redirect
-from django.http import HttpResponse
-from django.template import loader
-from .models import SourceChecklist, ChecklistTemplateItems
-from .forms import ChecklistTemplateForm, ChecklistItemForm
 from .views_checklist_template import *
 from .views_assign_checklist_to_user import *
-import datetime, enumerations
+import enumerations
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.models import User, Permission
+from brighterchecklist.shared.security_check import *
 
 @login_required
 def manager(request):
-    print("logged in user:", request.user.username, "\nid: ", request.user.id)
+    # print("logged in user:", request.user.username, "\nid: ", request.user.id)
+
     all_checklists = SourceChecklist.objects.all().filter(owner=request.user)
     template = loader.get_template('manager/checklist_manager_main.html')
 
@@ -35,9 +31,9 @@ def manager(request):
         'navigation': enumerations.Navigation.manager.name,
     }
 
-
     # https://stackoverflow.com/questions/30559020/django-login-template-doesnt-recognize-logged-user
     # return render(request, 'manager/checklist_manager_main.html', context)
+
     return HttpResponse(template.render(context, request))
 
 @login_required
@@ -45,6 +41,7 @@ def new(request):
     new_checklist = SourceChecklist()
     form = ChecklistTemplateForm()
     template = loader.get_template('manager/checklist_manager_entry.html')
+
     context = {
         'form': form,
         'navigation': enumerations.Navigation.manager.name,
@@ -55,7 +52,9 @@ def new(request):
 @login_required
 def delete(request, id):
     checklist_to_delete = SourceChecklist.objects.get(id=id)
-    checklist_to_delete.delete()
+
+    if check_security(checklist_to_delete.owner, request.user):
+        checklist_to_delete.delete()
 
     return redirect('/manager/')
 
@@ -64,19 +63,24 @@ def edit(request, id):
     checklist_data = SourceChecklist.objects.get(id=id)
     template = loader.get_template('manager/checklist_manager_entry.html')
 
-    ## Put the values in the form
-    data = {
-        'checklist_name':checklist_data.checklist_name,
-        'checklist_details':checklist_data.checklist_details
-    }
-    form = ChecklistTemplateForm(data)
-    # print (vars(form))
+    ## SECURITY: Check for owner.
+    if check_security(checklist_data.owner, request.user):
+        ## Put the values in the form
+        data = {
+            'checklist_name':checklist_data.checklist_name,
+            'checklist_details':checklist_data.checklist_details
+        }
+        form = ChecklistTemplateForm(data)
+        # print (vars(form))
 
-    context = {
-        'form': form,
-        'checklist_data': checklist_data,
-        'navigation': enumerations.Navigation.manager.name,
-    }
+        context = {
+            'form': form,
+            'checklist_data': checklist_data,
+            'navigation': enumerations.Navigation.manager.name,
+        }
+    else:
+        return redirect('/manager/')
+
 
     return HttpResponse(template.render(context, request))
 
@@ -87,10 +91,10 @@ def save(request, id):
         checklist_to_update = SourceChecklist(owner=request.user)
     else:
         checklist_to_update = SourceChecklist.objects.get(id=id)
-        print (checklist_to_update.owner, request.user, checklist_to_update.owner != request.user)
+        # print (checklist_to_update.owner, request.user, checklist_to_update.owner != request.user)
 
     ## SECURITY: Check for owner.
-    if checklist_to_update.owner == request.user:
+    if check_security(checklist_to_update.owner, request.user):
         form = ChecklistTemplateForm(request.POST)
 
         if form.is_valid():
